+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.5.0 (Next)
+### 1.5.0
 **GoSS**
 - Convert `flags` parameter in `validate_docker` from `list<string>` to `map` type.